requirements.txt
django-environ==0.11.2
requests == 2.31.0
